♻️ Refactor: Create GyazoAPIFactory object
